Fjerner noen dependencies. Skiller ut databaseService inn i  FavouritesService og Database. Renamer frontend til skribenten-bff i docker-compose.yml.
